requirements.txt
mysql-connector-python~=9.3.0
PyYAML~=6.0.2
Kivy~=2.3.1